use gridResolution instead of stepSize

+++ flink/src/main/java/org/apache/sedona/flink/expressions/Functions.java
@@ -52,7 +52,7 @@ public class Functions {
     @DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class)
     public Geometry eval(
         @DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class) Object o,
-        @DataTypeHint("Integer") Double gridResolution) {
+        @DataTypeHint("Integer") Integer gridResolution) {
       Geometry geom = (Geometry) o;
       return org.apache.sedona.common.Functions.labelPoint(geom, gridResolution);
     }
@@ -60,7 +60,7 @@ public class Functions {
     @DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class)
     public Geometry eval(
         @DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class) Object o,
-        @DataTypeHint("Integer") Double gridResolution,
+        @DataTypeHint("Integer") Integer gridResolution,
         @DataTypeHint("Double") Double goodnessThreshold) {
       Geometry geom = (Geometry) o;
       return org.apache.sedona.common.Functions.labelPoint(geom, gridResolution, goodnessThreshold);
